feature: add preview form view route, seperate view logic, consolidate render into viewModel, add and update unit tests
