+++ src/FormBuilder/ViewModels/FormBuilderViewModel.php
@@ -3,6 +3,8 @@
 namespace Give\FormBuilder\ViewModels;
 
 use Give\FormBuilder\ValueObjects\FormBuilderRestRouteConfig;
+use Give\NextGen\DonationForm\Actions\GenerateDonationFormPreviewRouteUrl;
+use Give\NextGen\DonationForm\Models\DonationForm;
 use Give\NextGen\Framework\FormDesigns\FormDesign;
 use Give\NextGen\Framework\FormDesigns\Registrars\FormDesignRegistrar;
 
@@ -13,12 +15,15 @@ class FormBuilderViewModel
      */
     public function storageData(int $donationFormId): array
     {
+        /** @var DonationForm $donationForm */
+        $donationForm = DonationForm::find($donationFormId);
+
         return [
             'resourceURL' => rest_url(FormBuilderRestRouteConfig::NAMESPACE . '/form/' . $donationFormId),
-            'previewURL' => site_url("?givewp-view=donation-form&form-id=$donationFormId"),
+            'previewURL' => (new GenerateDonationFormPreviewRouteUrl())($donationFormId),
             'nonce' => wp_create_nonce('wp_rest'),
-            'blockData' => get_post($donationFormId)->post_content,
-            'settings' => get_post_meta($donationFormId, 'formBuilderSettings', true),
+            'blockData' => $donationForm->blocks->toJson(),
+            'settings' => json_encode($donationForm->settings),
             'currency' => give_get_currency(),
             'formDesigns' => array_map(static function ($designClass) {
                 /** @var FormDesign $design */
