+++ src/NextGen/DonationForm/Actions/GenerateDonationFormPreviewRouteUrl.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Give\NextGen\DonationForm\Actions;
+
+
+/**
+ * @unreleased
+ */
+class GenerateDonationFormPreviewRouteUrl
+{
+    /**
+     * @unreleased
+     */
+    public function __invoke(int $formId): string
+    {
+        $args = [
+            'givewp-view' => 'donation-form-preview',
+            'form-id' => $formId
+        ];
+
+        return esc_url(
+            add_query_arg(
+                $args,
+                site_url()
+            )
+        );
+    }
+}
