+++ src/NextGen/DonationForm/Controllers/DonationFormViewController.php
@@ -2,29 +2,15 @@
 
 namespace Give\NextGen\DonationForm\Controllers;
 
-use Give\Framework\EnqueueScript;
+use Give\NextGen\DonationForm\DataTransferObjects\DonationFormPreviewRouteData;
 use Give\NextGen\DonationForm\DataTransferObjects\DonationFormViewRouteData;
 use Give\NextGen\DonationForm\Models\DonationForm;
-use Give\NextGen\DonationForm\Repositories\DonationFormRepository;
 use Give\NextGen\DonationForm\ViewModels\DonationFormViewModel;
-use Give\NextGen\Framework\FormDesigns\Registrars\FormDesignRegistrar;
 
 class DonationFormViewController
 {
     /**
      * This renders the donation form view.
-     *
-     * This is the order of loading:
-     * 1. Enqueue global styles from WP.
-     *  - This ensures template compatability with global WP css variables as needed. Loads before our templates, so they can use things like global font-family, etc.
-     * 2. Enqueue our donation form specific scripts & styles.
-     *  - We will let WP handle the actual printing depending on how they were enqueued.
-     * 3. Call the specific WP functions wp_print_styles() and wp_print_head_scripts()
-     *  - This will only print the styles and scripts that are enqueued within our route - so we don't have to dequeue a bunch of stuff.
-     * 4. Manually echo our window data and root div for our React app to consume
-     * 5. Finally, call the specific WP function wp_print_footer_scripts()
-     *  - This will only print the footer scripts that are enqueued within our route.
-     *
      *
      * @unreleased
      */
@@ -35,121 +21,29 @@ class DonationFormViewController
 
         $viewModel = new DonationFormViewModel(
             $donationForm->id,
+            $donationForm->blocks,
+            $donationForm->settings
+        );
+
+        return $viewModel->render();
+    }
+
+    /**
+     * This renders the donation form preview
+     *
+     * @unreleased
+     */
+    public function preview(DonationFormPreviewRouteData $data): string
+    {
+        /** @var DonationForm $donationForm */
+        $donationForm = DonationForm::find($data->formId);
+
+        $viewModel = new DonationFormViewModel(
+            $donationForm->id,
             $data->formBlocks ?: $donationForm->blocks,
             array_merge($donationForm->settings, $data->formSettings)
         );
 
-        wp_enqueue_global_styles();
-
-        $this->enqueueFormScripts(
-            $data->formId,
-            $viewModel->designId()
-        );
-
-        ob_start();
-        wp_print_styles();
-        wp_print_head_scripts();
-        ?>
-
-        <script>
-            window.giveNextGenExports = <?= wp_json_encode($viewModel->exports()) ?>;
-        </script>
-
-        <div
-            id="root-give-next-gen-donation-form-block"
-            class="givewp-donation-form-block"
-            style="
-                --give-primary-color:<?= $viewModel->primaryColor() ?>;
-                --give-secondary-color:<?= $viewModel->secondaryColor() ?>;
-                "
-        ></div>
-
-        <?php
-        wp_print_footer_scripts();
-        echo ob_get_clean();
-
-        exit();
-    }
-
-    /**
-     * Loads scripts in order: [Registrars, Designs, Gateways, Block]
-     *
-     * @unreleased
-     *
-     * @return void
-     */
-    private function enqueueFormScripts(int $formId, string $formDesignId)
-    {
-        /** @var DonationFormRepository $donationFormRepository */
-        $donationFormRepository = give(DonationFormRepository::class);
-
-        // load registrars
-        (new EnqueueScript(
-            'givewp-donation-form-registrars-js',
-            'build/donationFormRegistrars.js',
-            GIVE_NEXT_GEN_DIR,
-            GIVE_NEXT_GEN_URL,
-            'give'
-        ))->loadInFooter()->enqueue();
-
-        // load template
-        /** @var FormDesignRegistrar $formDesignRegistrar */
-        $formDesignRegistrar = give(FormDesignRegistrar::class);
-
-        // silently fail if design is missing for some reason
-        if ($formDesignRegistrar->hasDesign($formDesignId)) {
-            $design = $formDesignRegistrar->getDesign($formDesignId);
-
-            if ($design->css()) {
-                wp_enqueue_style('givewp-form-design-' . $design::id(), $design->css());
-            }
-
-            if ($design->js()) {
-                wp_enqueue_script(
-                    'givewp-form-design-' . $design::id(),
-                    $design->js(),
-                    array_merge(
-                        ['givewp-donation-form-registrars-js'],
-                        $design->dependencies()
-                    ),
-                    false,
-                    true
-                );
-            }
-        }
-
-        // load gateways
-        foreach ($donationFormRepository->getEnabledPaymentGateways($formId) as $gateway) {
-            if (method_exists($gateway, 'enqueueScript')) {
-                /** @var EnqueueScript $script */
-                $script = $gateway->enqueueScript();
-
-                $script->dependencies(['givewp-donation-form-registrars-js'])
-                    ->loadInFooter()
-                    ->enqueue();
-            }
-        }
-
-        // load block - since this is using render_callback viewScript in blocks.json will not work.
-        (new EnqueueScript(
-            'givewp-next-gen-donation-form-block-js',
-            'build/donationFormBlockApp.js',
-            GIVE_NEXT_GEN_DIR,
-            GIVE_NEXT_GEN_URL,
-            'give'
-        ))->dependencies(['givewp-donation-form-registrars-js'])->loadInFooter()->enqueue();
-
-        /**
-         * Load iframeResizer.contentWindow.min.js inside iframe
-         *
-         * @see https://github.com/davidjbradshaw/iframe-resizer
-         */
-        (new EnqueueScript(
-            'givewp-donation-form-embed-inside',
-            'build/donationFormEmbedInside.js',
-            GIVE_NEXT_GEN_DIR,
-            GIVE_NEXT_GEN_URL,
-            'give'
-        ))->loadInFooter()->enqueue();
+        return $viewModel->render();
     }
 }
