+++ src/NextGen/DonationForm/DataTransferObjects/DonationFormViewRouteData.php
@@ -2,8 +2,6 @@
 
 namespace Give\NextGen\DonationForm\DataTransferObjects;
 
-use Give\NextGen\Framework\Blocks\BlockCollection;
-
 /**
  * @unreleased
  */
@@ -15,19 +13,6 @@ class DonationFormViewRouteData
     public $formId;
 
     /**
-     * @var BlockCollection|null
-     */
-    public $formBlocks;
-
-    /**
-     * @var array
-     */
-    public $formSettings;
-
-    /**
-     * Convert data from request into DTO
-     *
-     * @param  array{form-id: string, form-settings: string, form-blocks: string}  $request
      *
      * @unreleased
      */
@@ -36,10 +21,6 @@ class DonationFormViewRouteData
         $self = new static();
 
         $self->formId = (int)$request['form-id'];
-        $self->formSettings = !empty($request['form-settings']) ? json_decode($request['form-settings'], true) : [];
-        $self->formBlocks = !empty($request['form-blocks']) ? BlockCollection::fromJson(
-            $request['form-blocks']
-        ) : null;
 
         return $self;
     }
