+++ tests/Unit/Actions/GenerateDonationFormPreviewRouteUrlTest.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace TestsNextGen\Unit\Actions;
+
+use Give\NextGen\DonationForm\Actions\GenerateDonationFormPreviewRouteUrl;
+use GiveTests\TestCase;
+
+/**
+ * @unreleased
+ */
+class GenerateDonationFormPreviewRouteUrlTest extends TestCase
+{
+    /**
+     * @unreleased
+     *
+     * @return void
+     */
+    public function testShouldReturnValidUrl()
+    {
+        $viewUrl = (new GenerateDonationFormPreviewRouteUrl())(1);
+
+        $this->assertSame(esc_url(
+            add_query_arg(
+                [
+                    'givewp-view' => 'donation-form-preview',
+                    'form-id' => 1
+                ],
+                site_url()
+            )
+        ), $viewUrl);
+    }
+}
