+++ tests/Unit/ViewModels/FormBuilderViewModelTest.php
@@ -5,6 +5,7 @@ namespace TestsNextGen\Unit\VieModels;
 use Exception;
 use Give\FormBuilder\ValueObjects\FormBuilderRestRouteConfig;
 use Give\FormBuilder\ViewModels\FormBuilderViewModel;
+use Give\NextGen\DonationForm\Actions\GenerateDonationFormPreviewRouteUrl;
 use Give\NextGen\DonationForm\Models\DonationForm;
 use GiveTests\TestCase;
 use GiveTests\TestTraits\RefreshDatabase;
@@ -29,7 +30,7 @@ class FormBuilderViewModelTest extends TestCase
         $this->assertArraySubset(
             [
                 'resourceURL' => rest_url(FormBuilderRestRouteConfig::NAMESPACE . '/form/' . $formId),
-                'previewURL' => site_url("?givewp-view=donation-form&form-id=$formId"),
+                'previewURL' => (new GenerateDonationFormPreviewRouteUrl())($formId),
                 'nonce' => wp_create_nonce('wp_rest'),
                 'blockData' => get_post($formId)->post_content,
                 'settings' => get_post_meta($formId, 'formBuilderSettings', true),
